css/css-fonts/parsing/font-variant-alternates-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Fonts Level 3: parsing font-variant-alternates with invalid values</title>
<link rel="author" title="Eric Willigers" href="mailto:ericwilligers@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-fonts/#font-variant-alternates-prop">
<meta name="assert" content="font-variant-alternates supports only the specified grammar.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("font-variant-alternates", "auto");
test_invalid_value("font-variant-alternates", "normal historical-forms");
test_invalid_value("font-variant-alternates", "styleset(name1) styleset(name2)");
</script>
</body>
</html>
